libnitrokey/unittest/requirements.txt
cffi
pytest-repeat
pytest-randomly
oath